Rename bar to chart.ts
