+++ src/notebooks/Controller.ts
@@ -6,7 +6,7 @@ import * as mdTable from 'json-to-markdown-table';
 import { getInstance } from '../base';
 import { CommandResult } from '@halcyontech/vscode-ibmi-types';
 import { JobManager } from '../config';
-import { ChartType, chartTypes, generateChart } from './logic/bar';
+import { ChartType, chartTypes, generateChart } from './logic/charts';
 
 export class IBMiController {
   readonly controllerId = `db2i-notebook-controller-id`;
